backend/requirements.txt
# Core FastAPI
fastapi==0.115.4
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1

# RAG & LLM
openai==1.54.4
pinecone-client==5.0.1
langchain==0.3.7
langchain-openai==0.2.8
langchain-community==0.3.7

# Document Processing (lightweight)
pypdf==5.1.0
python-multipart==0.0.17

# HTTP & Utils
httpx==0.27.2
tenacity==9.0.0

# CORS
python-jose[cryptography]==3.3.0